Fix for multi server deployment. Use md5 of the file contents instead of the file modification time.
This fixes the issue of deploying the code to multiple servers. Previously the hash of the file would be different on each server because the file modification time would be off by a few milliseconds. This causes 404's when the servers are load balanced and a request from server a is fulfilled by server b.